update scripts for LD map

+++ 03_filtering/imp/3_merge/3-0_merge_misc.sh
@@ -13,14 +13,10 @@ pgen_to_bed () {
     local cores=$4
     local plink_cmd=$5
     
-    if [ "${pvarfile%.zst}.zst" == "${pvarfile}" ] ; then
-        vzs="vzs"    
-    else
-        vzs=""
-    fi
-    local pfile=$(echo $zstfile | sed -e "s/.zst$//g" | sed -e "s/.pvar$//g")    
+    local pfile=$(echo ${zstfile%.zst} | sed -e "s/.pvar$//g")    
     
-    plink2 --memory ${plink_mem} --threads ${cores} --pfile "${pfile}" "${vzs}" --out "${bfile}" --${plink_cmd}
+    echo plink2 --memory ${plink_mem} --threads ${cores} --pfile "${pfile}" $([ "${pvarfile%.zst}.zst" == "${pvarfile}" ] && echo "vzs" || echo "") --out "${bfile}" --${plink_cmd}
+    plink2      --memory ${plink_mem} --threads ${cores} --pfile "${pfile}" $([ "${pvarfile%.zst}.zst" == "${pvarfile}" ] && echo "vzs" || echo "") --out "${bfile}" --${plink_cmd}
 }
 
 prep_files () {
@@ -37,25 +33,31 @@ prep_files () {
     local tmp_bed="${tmp_dir}/$(basename $pfile).bed"
     local tmp_fam="${tmp_dir}/$(basename $pfile).fam"
 
-    # generate bed/fam file from pfile when needed, otherwise, place a sym link.
-    if [ ! -f ${tmp_bed} ] ; then
+    echo ${tmp_bed}
+    echo ${tmp_bim}
+    echo ${tmp_fam}
+
+    # generate bed/fam file from pfile when needed.
+    # we will generate bim file later
+    if [ ! -s ${tmp_bed} ] ; then
+        # bed file does not exist --> generate bed + fam (+ bim) file
         pgen_to_bed ${zstfile} ${tmp_bed%.bed} ${plink_mem} ${cores} "make-bed"
         mv ${tmp_bed%.bed}.log ${tmp_bed%.bed}.bed.log
-    else
-        ln -s ${pfile}.bed ${tmp_bed}
-    fi
-    
-    if [ ! -f ${tmp_fam} ] ; then
+    elif [ ! -s ${tmp_fam} ] ; then
+        # fam file does not exist --> generate fam (+ bim) file
         pgen_to_bed ${zstfile} ${tmp_bed%.bed} ${plink_mem} ${cores} "make-just-fam"
         mv ${tmp_bed%.bed}.log ${tmp_bed%.bed}.fam.log
-    else
-        ln -s ${pfile}.fam ${tmp_fam}
     fi
-
-    cat_or_zstdcat ${zstfile} | grep -v '##' > ${tmp_pvar}
-    Rscript ${helper_R_script} ${tmp_pvar} ${prefix} ${pfile}.IDs.tsv
-    rm ${tmp_pvar}
     
+    # generate the short variant ID file
+    if [ ! -s ${pfile}.IDs.tsv ] ; then
+        cat_or_zstdcat ${zstfile} | grep -v '##' > ${tmp_pvar}
+        Rscript ${helper_R_script} ${tmp_pvar} ${prefix} ${pfile}.IDs.tsv
+        rm ${tmp_pvar}
+    fi
+    
+    # write to a bim file
     cat ${pfile}.IDs.tsv \
     | awk -v FS='\t' -v OFS='\t' '(NR>1){print $1, $6, 0, $2, $5, $4}' > ${tmp_bim}
 }
+
